+++ src/main/java/lotto/domain/LottoUtil.java
@@ -2,7 +2,7 @@ package lotto.domain;
 
 public final class LottoUtil {
 
-    public static double calculateSaveOfReturn(int purchasePrice, int prize) {
+    public static double calculateSaveOfReturn(int purchasePrice, long prize) {
         return prize * 1.0 / purchasePrice;
     }
 }
